Rename functions to encapsulate that we are moving raw run config to pydantic, add test case for nested enum config, reduce code duplication by adding new type of traversal context
